Refactor: Header 컴포넌트 수정
- Modal 폴더명 PopupModal로 변경
- Header에서 이용되는 Modal들은 createPortal을 이용하여 Body에 Modal을 Append하는 것이 아닌, Header 태그 내에 Append하는 방식으로 변경
