+++ src/components/Header/LongHeader.jsx
@@ -6,9 +6,7 @@ import { Button } from '@/components/Button';
 import { Input } from '@/components/Input';
 import { Image } from '@/components/Image';
 import { getImageSrc } from '@/utils/helpers';
-import InformationModal from './Modal/InformationModal/InformationModal';
-import NotificationModal from './Modal/NotificationModal/NotificationModal';
-import SidebarModal from './Modal/SidebarModal/SidebarModal';
+import { InformationModal, NotificationModal, SidebarModal } from './PopupModal';
 import MenuIcon from '@mui/icons-material/Menu';
 import SearchIcon from '@mui/icons-material/Search';
 import NotificationsIcon from '@mui/icons-material/Notifications';
